backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
anthropic==0.75.0
chromadb==0.4.15
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
cryptography==41.0.7
numpy==1.24.3
huggingface-hub==0.16.4

# PostgreSQL support
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# JWT for local authentication
djangorestframework-simplejwt==5.3.1
